requirements.txt
python-dotenv==0.9.1
urllib3==1.26.11
python-telegram-bot==13.13